backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
faiss-cpu==1.10.0
numpy==2.2.3
